+++ test/runtests.jl
@@ -1,6 +1,7 @@
 using Ferrite
 using Tensors
 using Test
+using Logging
 using ForwardDiff
 import SHA
 using Random
@@ -28,5 +29,6 @@ include("test_pointevaluation.jl")
 # include("test_notebooks.jl")
 include("test_apply_rhs.jl")
 include("test_apply_analytical.jl")
+HAS_EXTENSIONS && include("blockarrays.jl")
 include("test_examples.jl")
 @test all(x -> isdefined(Ferrite, x), names(Ferrite))  # Test that all exported symbols are defined